requirements.txt
streamlit==1.33.0
pandas==2.0.3
matplotlib==3.7.2
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4
tensorflow==2.15.0
google-generativeai==0.3.2
pygame==2.5.2